moved Navigation and Sidebar to Components folder
