+++ client/src/Components/Navigation/Navbar.js
@@ -1,9 +1,8 @@
 import { Link } from "react-router-dom"
-import {AuthenticationButton} from "../Features/Auth"
+import {AuthenticationButton} from "../../Features/Auth"
 import { useAuth0 } from "@auth0/auth0-react";
 import {useState} from 'react'
 import NavModal from './NavModal'
-
 
 export default function NavBar(){
   const [show, setShow] = useState(false)
