--- client/src/testdata.js
@@ -1,40 +1,0 @@
-export const recipeDataObj = [{
-    id: 29,
-    fileName: "/ingred.png.png.png",
-    location: "Ingredients",
-    OcrResult: {
-        data:{
-            text:"- 14 cup sugar\n//M\\\n2 pounds skinless catfish fillets\n//—_—___——_————_——_“\n5 garlic cloves\n///_——_——.—\na cUp vegetable oil\n\n% . | Fresh cilantro\n",
-        }
-    },
-    recipeEditorContent:{
-        blocks:[{ 
-        key:"99au1",
-        text:"1/4 cup sugar",
-        type :"unstyled",
-        depth:0,
-        inlineStyleRanges:[],
-        entityRanges:[],
-        data:{}
-        },
-        { 
-            key:"5ho98",
-            text:"2 pounds skinless catfish fillets",
-            type :"unstyled",
-            depth:0,
-            inlineStyleRanges:[],
-            entityRanges:[],
-            data:{}
-        },
-        { 
-            key:"7f06b",
-            text:"5 garlic cloves",
-            type :"unstyled",
-            depth:0,
-            inlineStyleRanges:[],
-            entityRanges:[],
-            data:{}
-        }
-    ]   
-    }
-}]
